Move simple network declarations to gene_simple_network.h
